+++ includes/admin/notices/class-wc-woomercadopago-notices.php
@@ -80,7 +80,6 @@ class WC_WooMercadoPago_Notices {
 		if (
 			( class_exists( 'WC_WooMercadoPago_Module' ) && WC_WooMercadoPago_Module::is_wc_new_version() )
 			&&
-			// @todo need fix Processing form data without nonce verification
 			// @codingStandardsIgnoreLine
 			( isset( $_GET['page'] ) &&	'wc-settings' === sanitize_key( $_GET['page'] )
 			)
@@ -134,7 +133,6 @@ class WC_WooMercadoPago_Notices {
 		if (
 			( class_exists( 'WC_WooMercadoPago_Module' ) && WC_WooMercadoPago_Module::is_wc_new_version() )
 			&&
-			// @todo need fix Processing form data without nonce verification
 			// @codingStandardsIgnoreLine
 			( isset( $_GET['page'] ) && 'wc-settings' === sanitize_key( $_GET['page'] )
 			)